рефакторинг фикстур
