+++ niffler_e_2_e_tests_python/tests_ui/presentation/authorization/main/profile/conftest.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Callable
 
 import pytest
 from configs import Configs
@@ -6,26 +6,14 @@ from tests_ui.pages.profile_page import ProfilePage
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
-    from tests_ui.pages.login_page import LoginPage
     from tests_ui.pages.main_page import MainPage
     from tests_ui.pages.presentation_page import PresentationPage
-    from utils.database import DB
-
-
-pytest_plugins = ('fixtures.helper_database', 'fixtures.helper_category')
 
 
 @pytest.fixture(scope='session')
 def profile_page(driver: 'Page') -> ProfilePage:
     """Получаем страницу Profile со всей логикой ее."""
     return ProfilePage(driver)
-
-
-@pytest.fixture(scope='class')
-def clear_spend_and_category_before(db_niffler_spend: 'DB') -> None:
-    """Чистим таблицу category и spend."""
-    db_niffler_spend.execute('delete from spend')
-    db_niffler_spend.execute('delete from category')
 
 
 @pytest.fixture
@@ -40,11 +28,13 @@ def goto_profile_if_you_logged_in(profile_page: ProfilePage) -> None:
 
 @pytest.fixture
 def goto_profile_if_you_not_logged_in(
-    login_page: 'LoginPage', presentation_page: 'PresentationPage', main_page: 'MainPage'
+    presentation_page: 'PresentationPage',
+    main_page: 'MainPage',
+    goto_login_page_and_log_in: Callable[[str, str], None],
 ) -> None:
     """Перейти на страницу main если не авторизован и находишься на разных местах сайта."""
     if presentation_page.driver.url != ProfilePage.url:
-        login_page.goto_login_page_and_log_in(Configs.TEST_USER, Configs.TEST_PASSWORD)
+        goto_login_page_and_log_in(Configs.TEST_USER, Configs.TEST_PASSWORD)
         main_page.click_profile_button()
