+++ niffler_e_2_e_tests_python/tests_ui/presentation/registration/conftest.py
@@ -6,12 +6,6 @@ from tests_ui.pages.register_page import RegisterPage
 if TYPE_CHECKING:
     from playwright.sync_api import Page
 
-pytest_plugins = (
-    'tests_ui.presentation.authorization.conftest',
-    'tests_ui.presentation.authorization.main.conftest',
-    'fixtures.helper_database',
-)
-
 
 @pytest.fixture(scope='session')
 def registration_page(driver: 'Page') -> RegisterPage:
