+++ niffler_e_2_e_tests_python/tests_ui/utils/prepare_user.py
@@ -1,12 +1,9 @@
 from http import HTTPStatus
-from typing import TYPE_CHECKING
 
 from configs import Configs
+from requests import Response
 from tests_api.clients_api.user import User
-
-if TYPE_CHECKING:
-    from requests import Response
-    from utils.database import DB
+from utils.database import DB
 
 
 def prepare_test_user(db_niffler_auth: 'DB') -> None:
